+++ docs/docs/changelog.md
@@ -6,6 +6,8 @@
 ## v0.0.3 (in development)
 - Bug fixes for handling of not found places
 - Fix requests dependency
+- Bug fix DCStatusError 500 - handle cases where an entity input is not valid 
+for the resolve endpoint.
 
 ## v0.0.2 (2025-06-18)
 - Set up GitHub Actions for CI/CD PYPI deployment